+++ tests/conftest.py
@@ -1,3 +1,5 @@
+import re
+
 import pytest
 from scootbot3.client import make_client
 from scootbot3.listener import listen
@@ -45,13 +47,13 @@ def test_listener(mocker, client, make_request):
 
     This fixture factory returns a function with signature:
 
-        test_listener(input: str, output: str) -> None
+        test_listener(input: str, output: re.Pattern) -> None
 
     The listener will be presented with input message text and we will assert
-    that it responds with the expected output text.
+    that its response matches the given output pattern.
     """
 
-    def _test_listener(input: str, output: str) -> None:
+    def _test_listener(input: str, expected_output: re.Pattern) -> None:
         # mock
         post_mock = mocker.patch.object(client.web_client, "chat_postMessage")
         make_response_mock = mocker.patch.object(
@@ -64,6 +66,7 @@ def test_listener(mocker, client, make_request):
 
         # check mocks
         make_response_mock.assert_called_once_with(envelope_id="test_envelope_id")
-        post_mock.assert_called_once_with(channel="test_channel", text=output)
+        assert post_mock.call_args.kwargs["channel"] == "test_channel"
+        assert expected_output.match(post_mock.call_args.kwargs["text"])
 
     return _test_listener